requirements.txt
accelerate==1.1.1
datasets==3.1.0
diffusers==0.31.0
evaluate==0.4.3
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.3
peft==0.13.2
Pillow==11.0.0
torch==2.5.1
torchmetrics==1.5.2
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.2
